Create main.c
This is where I tried to create a "user-mode" like environment for the individual to do their work from a system thread. Some functions include: ReadMemory(), AttachToProcess(), GetModuleBase(), SynthesizeMouse(), SynthesizeKeyboard() and get key/mouse state functions which work in an asynchronous manner. See the comments in main.c on functions and their purpose.